// src/models/types.ts
export interface AddonManifest {
    id: string;
    author: string,
    name: string;
    description: string;
    homepage_url: string;
    version: string;
    primary_type: string;
    exec: string;
    content_scripts?: ContentScript[];
    web_accessible_resources?: string[];
    manifest_version: number;
    enabled: boolean;
}


export interface GatewaySpecificSettings {
    webthings: WebthingsSettings;
}

export interface WebthingsSettings {
    exec?: string;
    strict_min_version: string;
    strict_max_version: string;
    primary_type: 'adapter' | 'notifier' | 'extension';
}

export interface Options {
    default?: Record<string, unknown>;
    schema: Record<string, unknown>;
}

interface ContentScript {
    css?: string[];
    js?: string[];
}